✅ move commander and config tests to client/tests
