+++ crates/client/src/commander.rs
@@ -536,7 +536,7 @@ impl Default for Commander {
     }
 }
 
-fn get_crash_dir_and_ext(
+pub fn get_crash_dir_and_ext(
     root: &str,
     target: &str,
     hfuzz_run_args: &str,
@@ -569,7 +569,7 @@ fn get_crash_dir_and_ext(
     (crash_path, extension)
 }
 
-fn get_cmd_option_value<'a>(
+pub fn get_cmd_option_value<'a>(
     hfuzz_run_args: impl Iterator<Item = &'a str>,
     short_opt: &str,
     long_opt: &str,
@@ -605,7 +605,7 @@ fn get_cmd_option_value<'a>(
     value
 }
 
-fn get_crash_files(
+pub fn get_crash_files(
     dir: &PathBuf,
     extension: &str,
 ) -> Result<Vec<PathBuf>, Box<dyn std::error::Error>> {
@@ -624,218 +624,4 @@ fn get_crash_files(
         })
         .collect::<Vec<_>>();
     Ok(paths)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    #[test]
-    fn test_cmd_options_parsing() {
-        let mut command = String::from("-Q -v --extension fuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q --extension fuzz -v");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q -e fuzz -v");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q --extension fuzz -v --extension ");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, None);
-
-        command = String::from("-Q --extension fuzz -v -e ");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, None);
-
-        let mut command = String::from("--extension buzz -e fuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q -v -e fuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q -v -efuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q -v --ext fuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, Some("fuzz".to_string()));
-
-        command = String::from("-Q -v --extfuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, None);
-
-        command = String::from("-Q -v --workspace");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "--ext");
-        assert_eq!(extension, None);
-
-        command = String::from("-Q -v -e");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "", "--ext");
-        assert_eq!(extension, None);
-
-        command = String::from("-Q -v --ext fuzz");
-        let args = command.split_whitespace();
-
-        let extension = get_cmd_option_value(args, "-e", "");
-        assert_eq!(extension, None);
-    }
-
-    #[test]
-    fn test_get_crash_dir_and_ext() {
-        pub const TARGET: &str = "fuzz_0";
-        pub const TEST_CRASH_PATH: &str = "/home/fuzz/test-crash-path";
-
-        const ROOT: &str = "/home/fuzz/";
-
-        let default_crash_path = std::path::Path::new(HFUZZ_WORKSPACE_DEFAULT).join(TARGET);
-        let env_specified_crash_path = std::path::Path::new(TEST_CRASH_PATH).join(TARGET);
-
-        // this is default behavior
-        let (crash_dir, ext) = get_crash_dir_and_ext(ROOT, TARGET, "", HFUZZ_WORKSPACE_DEFAULT);
-
-        assert_eq!(crash_dir, default_crash_path);
-        assert_eq!(&ext, "fuzz");
-
-        // behavior where path is specified within env variable HFUZZ_WORKSPACE, but not within -W HFUZZ_RUN_ARGS
-        let (crash_dir, ext) = get_crash_dir_and_ext(ROOT, TARGET, "-Q -e", TEST_CRASH_PATH);
-
-        assert_eq!(crash_dir, env_specified_crash_path);
-        assert_eq!(&ext, "fuzz");
-
-        // behavior as above
-        let (crash_dir, ext) = get_crash_dir_and_ext(ROOT, TARGET, "-Q -e crash", TEST_CRASH_PATH);
-
-        assert_eq!(crash_dir, env_specified_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // test absolute path
-        // HFUZZ_WORKSPACE has default value however -W is set
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q -W /home/crash -e crash",
-            HFUZZ_WORKSPACE_DEFAULT,
-        );
-
-        let expected_crash_path = std::path::Path::new("/home/crash");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // test absolute path
-        // HFUZZ_WORKSPACE is set and -W is also set
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q --crash /home/crash -e crash",
-            TEST_CRASH_PATH,
-        );
-        let expected_crash_path = std::path::Path::new("/home/crash");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // test absolute path
-        // HFUZZ_WORKSPACE is set and -W is also set
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q --crash /home/crash/foo/bar/dead/beef -e crash",
-            TEST_CRASH_PATH,
-        );
-
-        let expected_crash_path = std::path::Path::new("/home/crash/foo/bar/dead/beef");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // test relative path
-        // HFUZZ_WORKSPACE is set and -W is also set, this time with relative path
-        let (crash_dir, ext) =
-            get_crash_dir_and_ext(ROOT, TARGET, "-Q -W ../crash -e crash", TEST_CRASH_PATH);
-
-        let expected_crash_path = std::path::Path::new(ROOT).join("../crash");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // test relative path
-        // HFUZZ_WORKSPACE is set and -W is also set, this time with relative path
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q -W ../../dead/beef/crash -e crash",
-            TEST_CRASH_PATH,
-        );
-
-        let expected_crash_path = std::path::Path::new(ROOT).join("../../dead/beef/crash");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // test relative path
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q --crash ../crash -e crash",
-            HFUZZ_WORKSPACE_DEFAULT,
-        );
-
-        let expected_crash_path = std::path::Path::new(ROOT).join("../crash");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // crash directory has precedence before workspace option , which have precedence before
-        // HFUZZ_WORKSPACE
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q --crash ../bitcoin/to/the/moon -W /workspace -e crash",
-            TEST_CRASH_PATH,
-        );
-
-        let expected_crash_path = std::path::Path::new(ROOT).join("../bitcoin/to/the/moon");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-
-        // crash directory has precedence before workspace HFUZZ_WORKSPACE
-        let (crash_dir, ext) = get_crash_dir_and_ext(
-            ROOT,
-            TARGET,
-            "-Q --crash /home/crashes/we/like/solana -e crash",
-            TEST_CRASH_PATH,
-        );
-
-        // If path is specified as absolute, the join will replace whole path.
-        let expected_crash_path = std::path::Path::new("/home/crashes/we/like/solana");
-
-        // let expected_crash_path = root.join("/home/crashes/we/like/solana");
-        assert_eq!(crash_dir, expected_crash_path);
-        assert_eq!(&ext, "crash");
-    }
 }
